src/backend/requirements.txt
# Python Backend Dependencies for Pilgrims Window App

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2  # Async MongoDB driver

# WebSocket
websockets==12.0
python-socketio==5.10.0
aioredis==2.0.1  # For WebSocket scaling

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# QR Code Generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Data Visualization
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
pandas==2.1.4

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# CORS
fastapi-cors==0.0.6

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
python-json-logger==2.0.7

# SMS/Email (for emergency alerts)
twilio==8.11.0
sendgrid==6.11.0

# Geospatial (for location tracking)
geopy==2.4.1

# Rate Limiting
slowapi==0.1.9

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2